fix: 24826 Clean up `TreeView` related classes and methods
