+++ platform-sdk/swirlds-common/src/main/java/com/swirlds/common/merkle/synchronization/LearningSynchronizer.java
@@ -128,7 +128,7 @@ public class LearningSynchronizer {
             logger.warn(RECONNECT.getMarker(), "Interrupted while waiting for work group termination");
         } catch (final Throwable t) {
             logger.info(RECONNECT.getMarker(), "Caught exception while receiving tree", t);
-            throw t;
+            throw new RuntimeException(t);
         }
 
         if (interruptException != null || workGroup.hasExceptions()) {
