+++ platform-sdk/swirlds-common/src/main/java/com/swirlds/common/merkle/synchronization/views/TeacherTreeView.java
@@ -4,9 +4,6 @@ package com.swirlds.common.merkle.synchronization.views;
 import com.swirlds.base.time.Time;
 import com.swirlds.common.merkle.synchronization.streams.AsyncInputStream;
 import com.swirlds.common.merkle.synchronization.streams.AsyncOutputStream;
-import com.swirlds.common.merkle.synchronization.utility.MerkleSynchronizationException;
-import java.io.IOException;
-import org.hiero.base.io.streams.SerializableDataOutputStream;
 import org.hiero.consensus.concurrent.pool.StandardWorkGroup;
 
 /**
@@ -14,7 +11,7 @@ import org.hiero.consensus.concurrent.pool.StandardWorkGroup;
  * the tree by the teacher.
  *
  */
-public interface TeacherTreeView extends TeacherHandleQueue, TeacherResponseQueue, TeacherResponseTracker, TreeView {
+public interface TeacherTreeView extends AutoCloseable {
 
     /**
      * For this tree view, start all required reconnect tasks in the given work group. Teaching synchronizer
@@ -28,41 +25,4 @@ public interface TeacherTreeView extends TeacherHandleQueue, TeacherResponseQueu
      */
     void startTeacherTasks(
             final Time time, final StandardWorkGroup workGroup, final AsyncInputStream in, final AsyncOutputStream out);
-
-    /**
-     * Write data for a merkle leaf to the stream.
-     *
-     * @param out
-     * 		the output stream
-     * @param leafPath
-     * 		the merkle leaf path
-     * @throws IOException
-     * 		if an IO problem occurs
-     * @throws MerkleSynchronizationException
-     * 		if the node is not a leaf
-     */
-    void serializeLeaf(SerializableDataOutputStream out, Long leafPath) throws IOException;
-
-    /**
-     * Serialize data required to reconstruct an internal node. Should not contain any
-     * data about children, number of children, or any metadata (i.e. data that is not hashed).
-     *
-     * @param out
-     * 		the output stream
-     * @param internalPath
-     * 		the internal node path to serialize
-     * @throws IOException
-     * 		if a problem is encountered with the stream
-     */
-    void serializeInternal(SerializableDataOutputStream out, Long internalPath) throws IOException;
-
-    /**
-     * Serialize all child hashes for a given node into a stream. Serialized bytes must be
-     * identical to what out.writeSerializableList(hashesList, false, true) method writes.
-     *
-     * @param parentPath node path
-     * @param out The output stream
-     * @throws IOException If an I/O error occurred
-     */
-    void writeChildHashes(Long parentPath, SerializableDataOutputStream out) throws IOException;
 }
