+++ platform-sdk/swirlds-virtualmap/src/main/java/com/swirlds/virtualmap/internal/reconnect/LearnerPullVirtualTreeView.java
@@ -6,7 +6,6 @@ import static com.swirlds.logging.legacy.LogMarker.RECONNECT;
 import com.swirlds.common.merkle.synchronization.stats.ReconnectMapStats;
 import com.swirlds.common.merkle.synchronization.streams.AsyncInputStream;
 import com.swirlds.common.merkle.synchronization.streams.AsyncOutputStream;
-import com.swirlds.common.merkle.synchronization.task.ExpectedLesson;
 import com.swirlds.common.merkle.synchronization.utility.MerkleSynchronizationException;
 import com.swirlds.common.merkle.synchronization.views.LearnerTreeView;
 import com.swirlds.virtualmap.VirtualMap;
@@ -15,7 +14,6 @@ import com.swirlds.virtualmap.internal.Path;
 import com.swirlds.virtualmap.internal.RecordAccessor;
 import com.swirlds.virtualmap.internal.merkle.VirtualMapMetadata;
 import edu.umd.cs.findbugs.annotations.NonNull;
-import java.io.IOException;
 import java.util.Map;
 import java.util.Objects;
 import java.util.Queue;
@@ -29,7 +27,6 @@ import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.hiero.base.crypto.Cryptography;
 import org.hiero.base.crypto.Hash;
-import org.hiero.base.io.streams.SerializableDataInputStream;
 import org.hiero.consensus.concurrent.pool.StandardWorkGroup;
 import org.hiero.consensus.reconnect.config.ReconnectConfig;
 
@@ -42,9 +39,28 @@ import org.hiero.consensus.reconnect.config.ReconnectConfig;
  * <p>This implementation is supposed to work with {@link TeacherPullVirtualTreeView} on the
  * teacher side.
  */
-public final class LearnerPullVirtualTreeView extends VirtualTreeViewBase implements LearnerTreeView {
+public final class LearnerPullVirtualTreeView implements LearnerTreeView {
 
     private static final Logger logger = LogManager.getLogger(LearnerPullVirtualTreeView.class);
+
+    /**
+     * The root node that is involved in reconnect. This would be the saved state for the teacher, and
+     * the new root node into which things are being serialized for the learner.
+     */
+    private final VirtualMap map;
+
+    /**
+     * The state representing the original, unmodified tree on the learner. For simplicity, on the teacher,
+     * this is the same as {@link #reconnectState}. For the learner, it is the state of the detached, unmodified
+     * tree.
+     */
+    private final VirtualMapMetadata originalState;
+
+    /**
+     * The state representing the tree being reconnected. For the teacher, this corresponds to the saved state.
+     * For the learner, this is the state of the tree being serialized into.
+     */
+    private final VirtualMapMetadata reconnectState;
 
     /**
      * Reconnect configuration.
@@ -119,7 +135,9 @@ public final class LearnerPullVirtualTreeView extends VirtualTreeViewBase implem
             @NonNull final ReconnectNodeRemover nodeRemover,
             @NonNull final NodeTraversalOrder traversalOrder,
             @NonNull final ReconnectMapStats mapStats) {
-        super(map, originalState, reconnectState);
+        this.map = Objects.requireNonNull(map);
+        this.originalState = Objects.requireNonNull(originalState);
+        this.reconnectState = Objects.requireNonNull(reconnectState);
         this.reconnectConfig = reconnectConfig;
         this.originalRecords = Objects.requireNonNull(originalRecords);
         this.nodeRemover = nodeRemover;
@@ -295,69 +313,8 @@ public final class LearnerPullVirtualTreeView extends VirtualTreeViewBase implem
      * {@inheritDoc}
      */
     @Override
-    public void expectLessonFor(
-            final Long parentPath, final int childIndex, final Long originalPath, final boolean nodeAlreadyPresent) {
-        throw new UnsupportedOperationException("LearnerPullVirtualTreeView.expectLessonFor()");
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    @Override
-    public ExpectedLesson getNextExpectedLesson() {
-        throw new UnsupportedOperationException("LearnerPullVirtualTreeView.getNextExpectedLesson()");
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    @Override
-    public boolean hasNextExpectedLesson() {
-        throw new UnsupportedOperationException("LearnerPullVirtualTreeView.hasNextExpectedLesson()");
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    @Override
-    public Long deserializeLeaf(final SerializableDataInputStream in) throws IOException {
-        throw new UnsupportedOperationException("LearnerPullVirtualTreeView.deserializeLeaf()");
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    @Override
-    public Long deserializeInternal(final SerializableDataInputStream in) throws IOException {
-        throw new UnsupportedOperationException("LearnerPullVirtualTreeView.deserializeInternal()");
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    @Override
     public void close() {
         nodeRemover.allNodesReceived();
         map.endLearnerReconnect();
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    @Override
-    public void setChild(final Long parent, final int childIndex, final Long child) {
-        // No-op
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    @Override
-    public void recordHashStats(
-            @NonNull final ReconnectMapStats mapStats,
-            @NonNull final Long parent,
-            final int childIndex,
-            final boolean nodeAlreadyPresent) {
-        throw new UnsupportedOperationException("The Reconnect Pull Model records the hash stats elsewhere");
     }
 }
